third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-expected.txt

This is a testharness.js-based test.
[FAIL] Trusted Type violation report: sample for SVGScriptElement href assignment
  assert_equals: expected "SVGScriptElement href|about:blank" but got "SVGAnimatedString baseVal|about:blank"
[FAIL] Trusted Type violation report: sample for SVGScriptElement href assignment by setAttribute
  assert_equals: expected "SVGScriptElement href|about:blank" but got "Element setAttribute|about:blank"
Harness: the test ran to completion.
